Day 11 of 365 days of Python and ML
No Idea!

There is an array of n integers. There are also 2 disjoint sets,A and B, each containing integers. You like all the integers in set A and dislike all the integers in set B. Your initial happiness is 0. For each i integer in the array, if ,i belongs A you add 1 to your happiness. If i belongs to B, you add -1 to your happiness. Otherwise, your happiness does not change. Output your final happiness at the end.